Restructure plugin header and source files
