+++ include/plugin/interfaces/IComponent.hpp
